fix: 礼物记录中房间用户id类型修改；websocketserver修改参数为类参数； feat：礼物记录添加，附带投币记录，可以升级角色等级

+++ src/main/java/com/xinf/entity/GiftRecord.java
@@ -26,9 +26,9 @@ public class GiftRecord implements Serializable {
     @TableId(value = "serial_gift_id",type = IdType.AUTO)
     private Long serialGiftId;
 
-    private Integer userId;
+    private Long userId;
 
-    private Integer roomId;
+    private Long roomId;
 
     private Integer giftId;
